Refactor: #56 - Rename Command enum
* Command enum을 역할에 맞는 controller 팩키지로 이동
* 그에 따른 영향을 받은 BaseBallApp도 같이 추가
